User can see in-app notification permission on settings screen
